Issue 764: Schedule not rendered correct.
Attempt to fix by firing on document.ready only. Seems to be mostly
an issue with synchronising, which cannot be reproduced locally.
It is not the compression nor the production-data. Yet somehow the
JS throws an error on NIL. Which makes me assume the document is not
ready and the element that is being placed is not there yet.